+++ Assets/MoralisWeb3ApiSdk/Moralis/Moralis.Web3Api/CloudApi/DefiApi.cs
@@ -122,7 +122,7 @@ namespace Moralis.Web3Api.CloudApi
 			if (toBlock != null) postBody.Add("to_block", ApiClient.ParameterToString(toBlock));
 			if (toDate != null) postBody.Add("to_date", ApiClient.ParameterToString(toDate));
 			if (providerUrl != null) postBody.Add("provider_url", ApiClient.ParameterToString(providerUrl));
-			if(chain != null) postBody.Add("chain", ApiClient.ParameterToHex((long)chain));
+			postBody.Add("chain", ApiClient.ParameterToHex((long)chain));
 
 			// Authentication setting, if any
 			String[] authSettings = new String[] { "ApiKeyAuth" };
@@ -177,7 +177,7 @@ namespace Moralis.Web3Api.CloudApi
 			if (token1Address != null) postBody.Add("token1_address", ApiClient.ParameterToString(token1Address));
 			if (toBlock != null) postBody.Add("to_block", ApiClient.ParameterToString(toBlock));
 			if (toDate != null) postBody.Add("to_date", ApiClient.ParameterToString(toDate));
-			if(chain != null) postBody.Add("chain", ApiClient.ParameterToHex((long)chain));
+			postBody.Add("chain", ApiClient.ParameterToHex((long)chain));
 
 			// Authentication setting, if any
 			String[] authSettings = new String[] { "ApiKeyAuth" };