+++ Assets/MoralisWeb3ApiSdk/Moralis/Moralis.WebGL/Moralis.Web3Api/UniversalWebClient.cs
@@ -36,6 +36,7 @@ namespace Moralis.WebGL.Web3Api
         {
             "x-parse-application-id",
             "x-parse-installation-id",
+            "x-parse-session-token",
             "content-type"
         };
 
@@ -67,10 +68,12 @@ namespace Moralis.WebGL.Web3Api
             {
                 foreach (KeyValuePair<string, string> header in httpRequest.Headers)
                 {
-                    if (!String.IsNullOrWhiteSpace(header.Value) && allowedHeaders.Contains(header.Key.ToLower()))
+                    if (webRequest.GetRequestHeader(header.Key) != null) continue;
+
+                    if (!String.IsNullOrWhiteSpace(header.Value) &&
+                        allowedHeaders.Contains(header.Key.ToLower()))
                     {
                         webRequest.SetRequestHeader(header.Key, header.Value);
-                        Debug.Log($"Adding Header: {header.Key} value: {header.Value}");
                     }
                 }
             }