+++ Assets/MoralisWeb3ApiSdk/Moralis/MoralisDotNet/MoralisLiveQueryCallbacks.cs
@@ -1,9 +1,14 @@
-﻿using Moralis.Platform.Abstractions;
+﻿
+
+#if UNITY_WEBGL
+using Moralis.WebGL.Platform.Abstractions;
+using Moralis.WebGL.Platform.Objects;
+using Moralis.WebGL.Platform.Queries.Live;
+#else
+using Moralis.Platform.Abstractions;
 using Moralis.Platform.Objects;
 using Moralis.Platform.Queries.Live;
-using System;
-using System.Collections.Generic;
-using System.Text;
+#endif
 
 namespace Moralis
 {