+++ Assets/MoralisWeb3ApiSdk/Moralis/MoralisDotNet/Platform/Objects/MoralisUser.cs
@@ -1,14 +1,51 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Moralis.Platform.Abstractions;
 using Moralis.Platform.Utilities;
+using Newtonsoft.Json;
 
 namespace Moralis.Platform.Objects
 {
+    class MoralisUserJsonConvertor : JsonConverter
+    {
+        public override bool CanConvert(Type objectType)
+        {
+            return true;
+        }
+        public override bool CanRead
+        {
+            get { return false; }
+        }
+
+        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        {
+            throw new NotImplementedException("Unnecessary because CanRead is false. The type will skip the converter.");
+        }
+
+        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        {
+            if (value == null || !(value is MoralisUser))
+            {
+                serializer.Serialize(writer, null);
+                return;
+            }
+
+            Dictionary<string, object> user = ((MoralisUser)value).ToParameterDictionary();
+
+            serializer.Serialize(writer, user);
+        }
+    }
+
+
+    [JsonConverter(typeof(MoralisUserJsonConvertor))]
     public class MoralisUser : MoralisObject
     {
+        private bool isSaving = false;
+
         public MoralisUser()
         {
             this.ClassName = "_User";
@@ -20,20 +57,32 @@ namespace Moralis.Platform.Objects
             DateTime? createdAt = null,
             DateTime? updatedAt = null,
             MoralisAcl ACL = null,
-            string sessionToken = null) : base("_User", objectId, sessionToken, createdAt, updatedAt, ACL)
+            string sessionToken = null,
+            string ethAddress = null) : base("_User", objectId, sessionToken, createdAt, updatedAt, ACL)
         {
             this.username = userName;
             this.authData = authData ?? new Dictionary<string, IDictionary<string, object>>();
+
+            if (!String.IsNullOrEmpty(ethAddress))
+            {
+                this.ethAddress = ethAddress;
+                this.accounts = new string[1];
+                this.accounts[0] = ethAddress;
+            }
+            else
+            {
+                accounts = new string[0];
+            }
         }
 
         public string username;
 
         public IDictionary<string, IDictionary<string, object>> authData; 
 
-        //public string sessionToken;
         public string password;
         public string email;
-
+        public string ethAddress;
+        public string[] accounts;
 
         internal static IDictionary<string, IAuthenticationProvider> Authenticators { get; } = new Dictionary<string, IAuthenticationProvider> { };
     
@@ -44,6 +93,50 @@ namespace Moralis.Platform.Objects
 
         internal Task SignUpAsync(Task toAwait, CancellationToken cancellationToken) => throw new NotFiniteNumberException();
 
+        public Dictionary<string, object> ToParameterDictionary()
+        {
+            List<string> propertiesToSkip = new List<string>(new string[]{ "createdat", "sessiontoken" });
+            Dictionary<string, object> result = new Dictionary<string, object>();
+
+            // Use reflection to get all string properties 
+            // that have getters and setters
+            IEnumerable<PropertyInfo> properties = from p in this.GetType().GetProperties()
+                             where p.CanRead &&
+                                   p.CanWrite
+                             select p;
+
+            foreach (PropertyInfo property in properties)
+            {
+                // Not all properties should be included on save
+                if (isSaving && propertiesToSkip.Contains(property.Name.ToLower())) continue;
+
+                var value = property.GetValue(this);
+
+                if (value != null)
+                {
+                    result.Add(property.Name, value);
+                }
+            }
+
+            IEnumerable<FieldInfo> fields = from f in this.GetType().GetFields()
+                         where f.IsPublic
+                         select f;
+
+            foreach (FieldInfo field in fields)
+            {
+                // Not all properties should be included on save
+                if (isSaving && propertiesToSkip.Contains(field.Name.ToLower())) continue;
+
+                var value = field.GetValue(this);
+
+                if (value != null)
+                {
+                    result.Add(field.Name, value);
+                }
+            }
+
+            return result;
+        }
 
         /// <summary>
         /// Signs up a new user. This will create a new ParseUser on the server and will also persist the
@@ -51,6 +144,11 @@ namespace Moralis.Platform.Objects
         /// password must be set before calling SignUpAsync.
         /// </summary>
         public Task SignUpAsync() => SignUpAsync(CancellationToken.None);
+
+        public void SetSaving(bool val)
+        {
+            isSaving = val;
+        }
 
         /// <summary>
         /// Signs up a new user. This will create a new ParseUser on the server and will also persist the