+++ Assets/MoralisWeb3ApiSdk/Moralis/MoralisLoginTokenRequest.cs
@@ -25,7 +25,7 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  *  SOFTWARE.
  */
-namespace Assets.Scripts.Moralis
+namespace Moralis
 {
     /// <summary>
     /// Payload request object for a REST login request